# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "App API",
        "version": "0.1"
    },
    "paths": {
        "/api/donations/{donation}/contributions": {
            "post": {
                "tags": [
                    "Contributions"
                ],
                "summary": "Create a new contribution for a specific donation",
                "operationId": "eec198215994827593f3813ca2de857a",
                "parameters": [
                    {
                        "name": "donation",
                        "in": "path",
                        "description": "ID of the donation to contribute to",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "description": "Amount of the contribution",
                                        "type": "number",
                                        "format": "float",
                                        "example": "50.00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contribution created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "donation_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "amount": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "50.00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The amount field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found - donation not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Donation not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - user not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/donations": {
            "get": {
                "summary": "Retrieve a list of all donations",
                "operationId": "efd463fc838a5d0b3edcc89ecc8374ac",
                "responses": {
                    "200": {
                        "description": "List of donations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "donations": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new donation",
                "operationId": "a83d56537f533c01a1daf0abefdec913",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "target_amount"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Help the Needy"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Donate to help the less fortunate."
                                    },
                                    "target_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "5000.00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Donation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "donation": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - validation error"
                    }
                }
            }
        },
        "/api/donations/{donation_id}": {
            "get": {
                "summary": "Retrieve details of a specific donation",
                "operationId": "43a9d53da3d3f7670763846d38dad965",
                "parameters": [
                    {
                        "name": "donation_id",
                        "in": "path",
                        "description": "ID of the donation to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Donation details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "donation": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Donation not found"
                    }
                }
            }
        },
        "/api/donations/{donation_id}/status": {
            "patch": {
                "summary": "Update the status of a donation if the target amount is reached",
                "operationId": "812a6cb53e5f4b4987e44b4eb9fa68cd",
                "parameters": [
                    {
                        "name": "donation_id",
                        "in": "path",
                        "description": "ID of the donation to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Donation status updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "donation": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Donation not found"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "summary": "Register a new user",
                "operationId": "2f25885f3b6f61ca11724ec302d2e439",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "User already exists error"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "summary": "Authenticate a user and return a token",
                "operationId": "0fc9610184083ac6dddab2c63139225a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authentication successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid credentials"
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "summary": "Get the authenticated user",
                "operationId": "2130b7cce3aebae5b2f0aecd8f3cf0d3",
                "responses": {
                    "200": {
                        "description": "Authenticated user details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - user not authenticated"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Contributions",
            "description": "Contributions"
        }
    ]
}